recuperation,modifier et suppresion gestion event media

+++ eventhub-backend/src/main/java/org/example/eventhubbackend/services/EventService.java
@@ -2,10 +2,10 @@ package org.example.eventhubbackend.services;
 
 import lombok.RequiredArgsConstructor;
 import org.example.eventhubbackend.dto.EventRequest;
-import org.example.eventhubbackend.entity.Event;
-import org.example.eventhubbackend.entity.EventMedia;
-import org.example.eventhubbackend.entity.User;
+import org.example.eventhubbackend.dto.EventUpdateRequest;
+import org.example.eventhubbackend.entity.*;
 import org.example.eventhubbackend.repository.EventRepository;
+import org.example.eventhubbackend.repository.ReservationRepository;
 import org.example.eventhubbackend.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
@@ -16,6 +16,7 @@ import java.util.List;
 public class EventService {
     private final EventRepository eventRepository;
     private final UserRepository userRepository;
+    private final ReservationRepository reservationRepository;
 
     public Event createEvent(EventRequest request) {
 
@@ -48,12 +49,39 @@ public class EventService {
         return eventRepository.save(event);
     }
     public List<Event> getAllEvents() {
-        return eventRepository.findAll();
+        return eventRepository.findAllwithMediaAndOrganizer();
     }
-    public Event getEventById(Long id) {
-        return eventRepository.findById(id).orElseThrow(()->new RuntimeException("Event not found"));
+    public List<Event> getEventByOrganizer(Long organizer) {
+        return eventRepository.findByOrganizerId(organizer);
     }
     public void deleteEvent(Long id) {
-        eventRepository.deleteById(id);
+       List<Reservation> pendingReservation =
+               reservationRepository.findByEventIdAndStatus(id, ReservationStatus.PENDING);
+       if (pendingReservation.isEmpty()) {
+           throw new RuntimeException
+                   ("Impossible de supprimer: des reservations encours existent");
+
+       }
+       eventRepository.deleteById(id);
+
+    }
+    public Event updateEvent(Long id, EventUpdateRequest request,Long organizer) {
+        Event event=eventRepository.findById(id).
+                orElseThrow(() -> new RuntimeException("Evenement introuvable"));
+        // securité :seul l'organisateur peut le modifier
+        if (!event.getOrganizer().getId().equals(organizer)) {
+            throw new RuntimeException("Vous, vous n'etes pas autorisé à modifier cet event");
+        }
+        event.setTitle(request.getTitle());
+        event.setDescription(request.getDescription());
+        event.setStartDate(request.getStartDate());
+        event.setEndDate(request.getEndDate());
+        event.setLocation(request.getLocation());
+        event.setAddress(request.getAddress());
+        event.setCity(request.getCity());
+        event.setCapacity(request.getCapacity());
+        event.setCategory(request.getCategory());
+        return  eventRepository.save(event);
+
     }
 }
